feat: split the components
