+++ apps/landing-page/src/app/(landing-page)/constants/animations.ts
@@ -1,0 +1,3 @@
+export const initialAnimationDelay = 0.5
+export const animationDuration = 0.85
+export const animationDelayByChild = 0.1
